pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-form-filler"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.5",
    "openai==1.51.0",
    "httpx<0.28",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "telegramify-markdown==1.1.2",
]

[tool.setuptools.packages.find]
where = ["src"]
